providers: move `get_command_environment()` to providers.py (#3927)
`get_command_environment()` is a snapcraft-specific call, so it
is moved out of the craft-providers interface.
