+++ snapcraft/providers/_lxd.py
@@ -28,7 +28,7 @@ from snapcraft import utils
 
 from ._buildd import BASE_TO_BUILDD_IMAGE_ALIAS, SnapcraftBuilddBaseConfiguration
 from ._provider import Provider
-from .providers import get_instance_name
+from .providers import get_command_environment, get_instance_name
 
 logger = logging.getLogger(__name__)
 
@@ -99,7 +99,7 @@ class LXDProvider(Provider):
         """
         return lxd.LXDInstance(
             name=instance_name,
-            default_command_environment=self.get_command_environment(),
+            default_command_environment=get_command_environment(),
             project=self.lxd_project,
             remote=self.lxd_remote,
         )
@@ -143,7 +143,7 @@ class LXDProvider(Provider):
         except lxd.LXDError as error:
             raise ProviderError(str(error)) from error
 
-        environment = self.get_command_environment(
+        environment = get_command_environment(
             http_proxy=http_proxy, https_proxy=https_proxy
         )
