+++ snapcraft/providers/providers.py
@@ -16,7 +16,47 @@
 
 """Snapcraft-specific code to interface with craft-providers."""
 
+import os
 from pathlib import Path
+from typing import Dict, Optional
+
+from craft_providers import bases
+
+
+def get_command_environment(
+    http_proxy: Optional[str] = None, https_proxy: Optional[str] = None
+) -> Dict[str, Optional[str]]:
+    """Construct an environment needed to execute a command.
+
+    :param http_proxy: http proxy to add to environment
+    :param https_proxy: https proxy to add to environment
+
+    :return: Dictionary of environmental variables.
+    """
+    env = bases.buildd.default_command_environment()
+    env["SNAPCRAFT_MANAGED_MODE"] = "1"
+
+    # Pass-through host environment that target may need.
+    for env_key in [
+        "http_proxy",
+        "https_proxy",
+        "no_proxy",
+        "SNAPCRAFT_ENABLE_EXPERIMENTAL_EXTENSIONS",
+        "SNAPCRAFT_BUILD_FOR",
+        "SNAPCRAFT_BUILD_INFO",
+        "SNAPCRAFT_IMAGE_INFO",
+    ]:
+        if env_key in os.environ:
+            env[env_key] = os.environ[env_key]
+
+    # if http[s]_proxy was specified as an argument, then prioritize this proxy
+    # over the proxy from the host's environment.
+    if http_proxy:
+        env["http_proxy"] = http_proxy
+    if https_proxy:
+        env["https_proxy"] = https_proxy
+
+    return env
 
 
 def get_instance_name(
