+++ tests/unit/providers/test_multipass.py
@@ -95,14 +95,14 @@ class TestMultipassLaunchedEnvironment:
                 call(host_source=Path.home() / ".ssh", target=Path("/root/.ssh")),
             ]
 
-    def test_command_environment(self, mocker, new_dir, multipass_instance_mock):
+    def test_get_command_environment(self, mocker, new_dir, multipass_instance_mock):
         """Verify launched_environment calls get_command_environment."""
         mocker.patch(
             "snapcraft.providers._multipass.multipass.launch",
             return_value=multipass_instance_mock,
         )
-        mock_base_config = mocker.patch(
-            "snapcraft.providers._multipass.MultipassProvider.get_command_environment",
+        mock_get_command_environment = mocker.patch(
+            "snapcraft.providers._multipass.get_command_environment",
         )
 
         prov = providers.MultipassProvider()
@@ -117,6 +117,6 @@ class TestMultipassLaunchedEnvironment:
             http_proxy="test-http",
             https_proxy="test-https",
         ):
-            assert mock_base_config.mock_calls == [
+            assert mock_get_command_environment.mock_calls == [
                 call(http_proxy="test-http", https_proxy="test-https")
             ]
